Fixed a problem with adding custom PlayerLevels
Code was causing a out of range error when trying to call custom levels added by the mod api.  Adjusted the code so it is more robust and can dynamically add more levels to the table when registering new levels as well as when looking for what emblem to use.